+++ src/new-client/package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@next/env": "^14.2.5",
-    "@testing-library/user-event": "^14.5.2",
+    "appointments-booking-service": "file:",
     "next": "14.2.5",
     "nhsuk-frontend": "^8.3.0",
     "react": "^18",
@@ -26,6 +26,7 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^6.4.8",
     "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.12",
     "@types/node": "^20",
     "@types/react": "^18",
