+++ src/new-client/src/app/site/[site]/users/page.test.tsx
@@ -3,18 +3,22 @@ import { Role, User } from '@types';
 import UsersPage from './page';
 import { fetchRoles } from '../../../lib/roles';
 import { fetchUsers } from '../../../lib/users';
+import { getMockUserAssignments, mockRoles } from '../../../testing/data';
 
 jest.mock('../../../lib/roles');
 jest.mock('../../../lib/users');
 
 const fetchUsersMock = fetchUsers as jest.Mock<Promise<User[]>>;
 const fetchRolesMock = fetchRoles as jest.Mock<Promise<Role[]>>;
+const mockSiteId = 'TEST';
 
 describe('<UserManagement />', () => {
-  it('renders', async () => {
-    fetchUsersMock.mockResolvedValue(mockUserAssignments);
+  beforeEach(() => {
+    fetchUsersMock.mockResolvedValue(getMockUserAssignments(mockSiteId));
     fetchRolesMock.mockResolvedValue(mockRoles);
+  });
 
+  it('renders', async () => {
     const jsx = await UsersPage({ params: { site: mockSiteId } });
     render(jsx);
 
@@ -26,9 +30,6 @@ describe('<UserManagement />', () => {
   });
 
   it('displays each user in the table', async () => {
-    fetchUsersMock.mockResolvedValue(mockUserAssignments);
-    fetchRolesMock.mockResolvedValue(mockRoles);
-
     const jsx = await UsersPage({ params: { site: mockSiteId } });
     await render(jsx);
 
@@ -51,39 +52,3 @@ describe('<UserManagement />', () => {
     expect(screen.getByRole('cell', { name: 'Role 3' })).toBeInTheDocument();
   });
 });
-
-const mockSiteId = '1000';
-const mockUserAssignments: User[] = [
-  {
-    id: 'test.one@nhs.net',
-    roleAssignments: [
-      { role: 'role-1', scope: `site:${mockSiteId}` },
-      {
-        role: 'role-2',
-        scope: `site:${mockSiteId}`,
-      },
-    ],
-  },
-  {
-    id: 'test.two@nhs.net',
-    roleAssignments: [{ role: 'role-3', scope: `site:${mockSiteId}` }],
-  },
-];
-
-const mockRoles: Role[] = [
-  {
-    displayName: 'Role 1',
-    id: 'role-1',
-    description: 'This is a short description of role 1.',
-  },
-  {
-    displayName: 'Role 2',
-    id: 'role-2',
-    description: 'This is a short description of role 2.',
-  },
-  {
-    displayName: 'Role 3',
-    id: 'role-3',
-    description: 'This is a short description of role 3.',
-  },
-];
